refactor(back): Update config path sequelize remode file useless
