--- firebase/configFirebase.json
@@ -1,10 +1,0 @@
-{
-    "apiKey": "AIzaSyCbVXVWjGxyqRGZMGb3SFCvWw0oNOu0S10",
-    "authDomain": "memoriz-f62b4.firebaseapp.com",
-    "databaseURL": "https://memoriz-f62b4.firebaseio.com",
-    "projectId": "memoriz-f62b4",
-    "storageBucket": "memoriz-f62b4.appspot.com",
-    "messagingSenderId": "233440240512"
-}
-
-
