--- fixtures/users.mock.js
@@ -1,19 +1,0 @@
-const ticketMocks = [
-  {
-    id: 1,
-    firstName: "kevin",
-    lastName: "mouga",
-  },
-  {
-    id: 2,
-    firstName: "Arnaud",
-    lastName: "Adon",
-  },
-  {
-    id: 3,
-    firstName: "Warren",
-    lastName: "Dong Tam",
-  }
-];
-
-module.exports = ticketMocks;
